backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
discord.py>=2.4.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite==0.20.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
